+++ src/services/extractDescription.ts
@@ -5,12 +5,19 @@ export function extractDescription(
     node: ComponentNode
 ): string | undefined {
 
-    if (Node.isFunctionDeclaration(node)) {
+    if (
+        Node.isClassDeclaration(node) ||
+        Node.isFunctionDeclaration(node)
+    ) {
         const docs = node.getJsDocs();
 
         return docs.length
             ? docs.map(doc => doc.getDescription().trim()).join("\n")
             : undefined;
+    }
+
+    if (Node.isExportAssignment(node)) {
+        return undefined;
     }
 
     const statement = node.getVariableStatement();