+++ src/types/ComponentNode.ts
@@ -1,5 +1,12 @@
-import { FunctionDeclaration, VariableDeclaration } from "ts-morph";
+import {
+    ClassDeclaration,
+    ExportAssignment,
+    FunctionDeclaration,
+    VariableDeclaration,
+} from "ts-morph";
 
 export type ComponentNode =
+    | ClassDeclaration
+    | ExportAssignment
     | FunctionDeclaration
     | VariableDeclaration;