+++ src/types/ScanOptions.ts
@@ -1,4 +1,5 @@
 export interface ScanOptions {
     include?: string[];
     exclude?: string[];
+    componentWrappers?: string[];
 }